// ler/src/modules/grpc/proto/ler.pb.ts
export interface GetCityTiersRequest {
  name: string;
}

export interface GetCityTiersResponse {
  name: string;
  sellerTier: number;
  buyerTier: number;
}

export interface MapLogisticsServicesRequest {
  sellerCityTier: number;
  buyerCityTier: number;
  isKeySeller: boolean;
}

export interface MapLogisticsServicesResponse {
  logisticServices: string;
  vendorId: string;
  isAvailableToPickup: boolean;
  serviceId: string;
  isAvailableToOneService: boolean;
}

export interface GetLogisticServicesRequest {
}

export interface GetLogisticServicesResponse {
  services: GetLogisticServicesResponse_LogisticService[];
  vendors: GetLogisticServicesResponse_LogisticVendor[];
}

export interface GetLogisticServicesResponse_LogisticService {
  serviceId: string;
  serviceName: string;
}

export interface GetLogisticServicesResponse_LogisticVendor {
  vendorId: string;
  vendorName: string;
  services: string;
}

export interface CreatePickupRequest {
  referenceNo: string;
  originCity: string;
  destinationCity: string;
  senderName: string;
  senderPhone: string;
  senderAddress: string;
  receiverName: string;
  receiverPhone: string;
  receiverAddress: string;
  trackingNumber: string;
  description: string;
}

export interface CreateShipmentReq {
  sender: User | undefined;
  receiver: User | undefined;
  trackingNumber: string;
  description: string;
  grandTotal: number;
  shipmentType: string;
  serviceName: string;
  isConsignment?: boolean | undefined;
}

export interface User {
  name: string;
  mobileNumber: string;
  address: string;
  email: string;
  userType: string;
  city: string;
  latitude?: string | undefined;
  longitude?: string | undefined;
}

export interface CreatePickupForAccessoriesRequest {
  referenceNo: string;
  originCity: string;
  destinationCity: string;
  senderName: string;
  senderPhone: string;
  senderAddress: string;
  receiverName: string;
  receiverPhone: string;
  receiverAddress: string;
  trackingNumber: string;
  description: string;
  skudetails: SkuDetails[];
}

export interface CreatePickupResponse {
  awbNo: string;
}

export interface CreateShipmentResponse {
  trackingNumber: string;
}

export interface GetPickupStatusRequest {
  awbNo: string[];
  isDelivered: boolean;
}

export interface SkuDetails {
  sku: string;
  description: string;
  cod: string;
  piece: string;
  weight: string;
}

export interface CancelShipmentRequest {
  /** Field for order or tracking ID */
  trackingOrOrderId: string;
}

export interface CancelShipmentResponse {
  success: boolean;
  message: string;
}
// ...
export const LER_SERVICE_NAME = "LerService";